fix: incorrect test cast

+++ ksqldb-engine/src/test/java/io/confluent/ksql/function/udf/math/TruncTest.java
@@ -214,7 +214,7 @@ public class TruncTest {
 
   @Test
   public void shoulldHandleNullValues() {
-    assertThat(udf.trunc((Integer) null), is((Integer) null));
+    assertThat(udf.trunc((Integer) null), is((Long) null));
     assertThat(udf.trunc((Long) null), is((Long) null));
     assertThat(udf.trunc((Double) null), is((Long) null));
     assertThat(udf.trunc((Double) null), is((Long) null));
